+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.4.14 (Unreleased)
 - [Enhancement] Improve low-level critical error reporting.
 - [Enhancement] Expand Kubernetes Liveness state reporting with critical errors detection.
-- [Enhancement] Save several string allocations and one array allocation on each job execution. when using Datadog instrumentation.
+- [Enhancement] Save several string allocations and one array allocation on each job execution when using Datadog instrumentation.
 - [Enhancement] Support `eofed` jobs in the AppSignal instrumentation.
 - [Enhancement] Allow running bootfile-less Rails setup Karafka CLI commands where stuff is configured in the initializers.
 - [Fix] `Instrumentation::Vendors::Datadog::LoggerListener` treats eof jobs as consume jobs.